// types.ts
export interface WorldPreset {
  id: string;
  name: string;
  description: string;
  thumbnail?: string;
  state: any; // The session state object
}

export interface MapPreset {
  id: string;
  name: string;
  description: string;
  planets: WorldPreset[];
}